qc/illumina_qc: move 'determine_qc_protocol' from utils to here, and update other code to accommodate this move. Also put the list of QC protocol names into a constant which can be accessed by other code.
